refactor: test, 모달 정보 상수 공유모듈로 분리
