--- apps/web/e2e/project-search/flows.ts
@@ -1,5 +1,0 @@
-import { Page } from '@playwright/test';
-
-export async function openModal(page: Page) {}
-export async function reachStep2(page: Page) {}
-export async function reachStep3(page: Page) {}
